feat: se agrega servicio de producto, get, getAll y create

+++ src/main/java/com/project/models/Product.java
@@ -16,7 +16,7 @@ public class Product {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_product")
-    private long id;
+    private int id;
 
     @Column(name = "name")
     private String name;
